+++ core/docs/changelog.md
@@ -1,7 +1,16 @@
 # TBD
+### Features
+* Added metrics client to track module's events (e.g.: when users load a module)
+
+### Breaking Changes
+* Change the `ApiContainerLauncher.LaunchWithDefaultVersion()` and `ApiContainerLauncher.LaunchWithCustomVersion()` methods API, added two new arguments `metricsUserID` and `didUserAcceptSendingMetrics`
+  * Users should add these two new arguments in every call
+* Change `ApiContainerService` constructor, now receives a new extra argument `metricsClient`
+  * Users should add this new argument in every call
+* Change `ApiContainerArgs` constructor, added two new arguments `metricsUserID` and `didUserAcceptSendingMetrics`
+  * Users should add these two new arguments in every call
 
 # 1.36.12
-
 ### Fixes
 * Fix `SoftPartitionConnection` class constructor bug upon `isValidPacketLossValue` value decision.